Added small change to MD_Range kernels

+++ src/Aeras/evaluators/Aeras_DOFGradInterpolationLevels_Def.hpp
@@ -134,7 +134,7 @@ evaluateFields(typename Traits::EvalData workset)
 #else
 #if defined(PHX_KOKKOS_DEVICE_TYPE_CUDA)
   DOFGradInterpolationLevels_Policy range(
-                {0,0,0}, {workset.numCells,numQPs,numLevels}, {256,1,1} );
+                {0,0,0}, {(int)workset.numCells,(int)numQPs,(int)numLevels}, {256,1,1} );
 #else
   DOFGradInterpolationLevels_Policy range ({(int)workset.numCells,
 		(int)numQPs,(int)numLevels});
@@ -232,7 +232,7 @@ evaluateFields(typename Traits::EvalData workset)
 #else
 #if defined(PHX_KOKKOS_DEVICE_TYPE_CUDA)
   DOFGradInterpolationLevels_noDeriv_Policy range(
-                {0,0,0}, {workset.numCells,numQPs,numLevels}, {256,1,1} );
+                {0,0,0}, {(int)workset.numCells,(int)numQPs,(int)numLevels}, {256,1,1} );
 #else 
   DOFGradInterpolationLevels_noDeriv_Policy range ({(int)workset.numCells,
                 (int)numQPs,(int)numLevels});
