+++ src/Aeras/evaluators/Aeras_XZHydrostatic_KineticEnergy_Def.hpp
@@ -90,7 +90,7 @@ evaluateFields(typename Traits::EvalData workset)
 #else
 #if defined(PHX_KOKKOS_DEVICE_TYPE_CUDA)
   XZHydrostatic_KineticEnergy_Policy range(
-                {0,0,0}, {workset.numCells,numNodes,numLevels}, {256,1,1} );
+                {0,0,0}, {(int)workset.numCells,(int)numNodes,(int)numLevels}, {256,1,1} );
 #else
   XZHydrostatic_KineticEnergy_Policy  range ({(int)workset.numCells,(int)numNodes,(int)numLevels});
 #endif
